Create Endsort
This is a little algorithm I made when I was bored at my desk job. It's not exactly the most efficient (O(n^2) if I'm not mistaken), but I just wanted some practice in JS to get used to the syntax.  I called in Endsort because it goes through the array and puts the largest value in the last array position and repeats this n times, while leaving the sorted partition untouched.